UI Updates and added logic
Discovered that layout types could be nested! This made playing with the slider and the two new buttons for it much easier.

Made the slider default to its horizontal state for the time being since it completely messes up the axis.

Added in logic to every button to observe proper communication with the viewmodel. Using Commands since they offer runtime modification. Could be useful in the future, and looks neat and organized.

Added an animation to the powerbutton frame from an example from the internet.

Changed listview to picker. Added in a list with the first UI type labeled "Phase I". If we use this program for future versions of our Rover, the UI will likely update. Instead of creating a new UI for each phase, we can keep each version available in our application. This will likely result in multiple viewmodels. We can use this in case we need to switch from one version to another without installing specific versions of the app.