Environment Updates, Name Changes
Added another environment type for KestrelHosting and made it so the program only updates the default hosting if in not in a development environment.

Changed all the names from the default Greeter project for Grpc to the Controller name for our rover project.

Renamed all files and folders as necessary.