Added Dependency Injection for configuration
app.settings was hardcoded. There may have been a way for it to accept new configurations, but it was easier to just add dependency injection. Using this model also allows me to added more robust logging in the future.

Added 2 message boxes for network-related issues that may occur.

Added startup logging capability. For example, if the program fails to startup due to missing appsettings, a log file will be created in the directory the program is running from.

Fixed issue where the single file app contained the cert and appsettings files.