New motors, new arduino code; notes on power issue
The front wheel motors were replaced with the same type as the back wheels, so now all motors can be given the same exact speed. The back wheels still shut off when the power goes low, and the front ones still continue. This is likely a power issue with the motor driver board itself. They work fine at faster speeds.

Before, we noted that the arduino was back propagating power into the pi. I suspected, but did not convey properly, that the motor driver board (which receives power seperately, was feeding into the arduino which then fed into the pi.

We troubleshooted and discoverred this was true. In the time since we last looked at it, dad created a new power design which is no longer necessary. The motor driver board has a simple jumper that ron discovered. Removing this jumper prevents the back-feeding of power.