Created new Rover Mobile Application
It's been quite a while since I last looked at the failed Rover Application. The app was in a broken state due to changes to Xamarin. Instead, I moved the Controller logic to a brand new Xamarin Forms Template.

Xamarin has improved tremendously in that time and was easy to get the logic integrated. From here, all I now need to do is implement the logic for GRPC that the windows application has and see if it works.