Finished Mobile UI
The new Xamarin application's UI should now be finished. All the connecting logic is set up, now all I need to do is add the background functionality and see if that works.

Since we have the template's item view, we may be able to modify it to hold/save IP addresses. The hard part will likely getting all the permissions set up for each platform... if GRPC is even supported.