Grpc works!
It took a lot of research and playing around in order to learn both Arduino Emulator network details as well as how to get Xamarin to use Grpc... as well as configure the server to support the method Xamarin requires.

There is a special version of Grpc that enables it to work through some Microsoft black magic and http header manipulation. It's not quite the super secure Http2, but it's good enough.

Added a new project to the Xamarin solution to Abstract the gRPC logic.

Will need to create a simple way to take in an IP address on command.

No longer requiring certificates for the moment.

Updated GrpcServer packages.