Local Network Scanning
Based on a few assumptions, I have basic network scanning functionality that should allow the app to pick up useable servers on the network.

I've only tested it on the emulator so far, but it works. I think there are some limitations/logic issues when scalling past class C networks, but it should work for the most part. Really, the biggest issue is how slow it will be for larger than class C networks.

One assumption is that the default gateway is always the first address. This is literally just a tool for our own use, and will likely always be the case... but still.

Another assumption is that we're using the wlan interface. I haven't seen many android devices, but I suspect that will almost always be the case.