Optimizations
After playing with it further, I decided to make the IP scanner the function that's called from the view model directly, and made it an async ienumerable so we can simply yield values back to the caller as we get them.

By yielding status messages back, I can both relate information to the UI as well as supply the ID of the last connectable Item so the UI can fetch that item by ID.

There's definitely improvements to be had, but it works pretty well. The largest hiccup was the fact that I needed to do a Task.WhenAny call in the client connection test due to either hardware or software limitations on the actual Google Pixel 2 device I own. Even though I had a deadline set, the connection to a random device on the network would get stuck. The WhenAny logic works fantastically by allowing a sort OR functionality between the tasks.

This logic is now fully functioning on my test device.